SwimDataProject/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
lxml==4.9.3
webdriver-manager==4.0.1
PyPDF2==3.0.1

# Data processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
alembic==1.12.0

# API
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
PyYAML==6.0.1
schedule==1.2.0
tqdm==4.66.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0
